requirements.txt
Flask==3.1.2
sqlalchemy==2.0.43
PyJWT==2.10.1
requests==2.32.5
gunicorn==23.0.0

